Make bitcoin_my_transactions_signing_keys_p() actually attempt signing
Rather than trying to do our own privkey accounting (which is going
to break as soon as we allow the user to specify privkeys which
might not have specific inputs associated to them), we simply sign
the transaction and check if it's complete. bitcoind kindly gives
us the complete field in the response of signrawtransaction, so this
is easy :).